tests/wpt/tests/custom-elements/upgrading/Document-importNode-customized-builtins.html

<!DOCTYPE html>
<link rel="help" href="https://dom.spec.whatwg.org/#dom-document-importnode">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/custom-elements-helpers.js"></script>
<body>
<script>
test_with_window((w, doc) => {
  class MyDiv extends HTMLDivElement {}
  class MyDiv2 extends w.HTMLDivElement {}
  customElements.define('my-div', MyDiv, { extends: 'div' });
  w.customElements.define('my-div', MyDiv2, { extends: 'div' });

  let original = document.createElement('div', { is: 'my-div' });
  assert_true(original instanceof MyDiv);

  let imported = doc.importNode(original);
  assert_true(imported instanceof MyDiv2);
}, 'built-in: document.importNode() should import custom elements successfully');

test_with_window((w, doc) => {
  class MyDiv2 extends w.HTMLDivElement {}
  w.customElements.define('my-div2', MyDiv2, { extends: 'div' });

  let original = document.createElement('div', { is: 'my-div2' });
  assert_equals(original.constructor, HTMLDivElement);

  let imported = doc.importNode(original);
  assert_true(imported instanceof MyDiv2);
}, 'built-in: document.importNode() should import "undefined" custom elements successfully');
</script>
</body>
